requirements.txt
selenium==4.23.1
pytest==8.3.2
requests
allure-pytest==2.13.5
allure-python-commons==2.13.5


